Refactor timeout audit log handling into class
